List of broken recipes for python3crystax, refs #1402
Populated the list of broken recipes and moved to a dedicated
`constants.py` file.
Broken recipes are mostly divided in 3 categories:
1) recipes not yet migrated from python2 to python3crystax
2) recipes that don't have environment var well setup
3) recipes that are not broken but require a system package
